chore(): refactor ws events into organised dir
