+++ src/types/index.ts
@@ -1,6 +1,5 @@
 export * from './response';
 export * from './request';
 export * from './shared';
-export * from './v5-shared';
+export * from './shared-v5';
 export * from './websockets';
-export * from './websocket.events';
