+++ src/types/response/v5-trade.ts
@@ -10,7 +10,7 @@ import {
   OrderTypeV5,
   PositionIdx,
   StopOrderTypeV5,
-} from '../v5-shared';
+} from '../shared-v5';
 
 export interface OrderResultV5 {
   orderId: string;
