+++ src/types/websockets/index.ts
@@ -1,0 +1,2 @@
+export * from './ws-general';
+export * from './ws-events';
