--- src/types/ws-events/failed-topic-subscription-confirmation.ts
@@ -1,6 +1,0 @@
-import { WebsocketTopicSubscriptionConfirmationEvent } from './topic-subscription-confirmation';
-
-export interface WebsocketFailedTopicSubscriptionConfirmationEvent
-  extends WebsocketTopicSubscriptionConfirmationEvent {
-  success: false;
-}
