--- src/types/ws-events/succeeded-topic-subscription-confirmation.ts
@@ -1,6 +1,0 @@
-import { WebsocketTopicSubscriptionConfirmationEvent } from './topic-subscription-confirmation';
-
-export interface WebsocketSucceededTopicSubscriptionConfirmationEvent
-  extends WebsocketTopicSubscriptionConfirmationEvent {
-  success: true;
-}
