--- src/types/ws-events/topic-subscription-confirmation.ts
@@ -1,7 +1,0 @@
-export interface WebsocketTopicSubscriptionConfirmationEvent {
-  op: 'subscribe';
-  req_id: string;
-  conn_id: string;
-  ret_msg: string;
-  success: boolean;
-}
